Separate precondition and state change function were joined.
The original design was to have a pair of precondition (upon the `BoardState`)
and a function that changes this board state (when precondition holds).
That is, to have a pair of `Ard`'s type parameters, somewhat like this:
  `(statePrecondition : BoardState -> Type)`
  `(stateChanges : (inS : BoardState) -> statePrecondition inS => BoardState)`
Unfortunately, this did not typecheck with Idris 1 for unknown reason.

That's why `Maybe`-based approach with monadic composition of joined
precondition and state changing function was applied.