`Ard`'s preconditions and effects tracking was redesigned again.
Now, we are back to the idea of separate precondition and description of
what's going on. Precondition is simply a function from the state
(facts list, at the moment) to `Type`. Since state is not a list of facts,
state change now is just another list of facts means to be appended
(in the sense of relative events origin) to the original one.