Postponing was redesigned and rebranded (now it's called `Fence`).
This, actually, is the original design, but somewhy firstly implemented
another way. The idea is that only one postponed (waiting for parallel)
action can be awakened after some event, because only one action can be
stored as a waiting for parallel at a time. Thus, each event can store
only one postponed action. But this postponed action may, however, have
its own actions waiting for it. This, postponing data structure is a
kind of a customized list with `Sync` values.

For more information, look at how `Sequential` of `Cooperative` and `f`
are scheduled: `currFence` of the whole `Sequential` is an action to be
run after `f`, not the `Cooperative`. So, when storing `f` as a
postponed (waiting for both sides of this `Cooperative`) we need to
store `currFence` in the continuation containing `f`. Thus, we need a
list-like data structure, which I called `Fence`.